prod-version/requirements.txt
# Core dependencies
requests>=2.31.0
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0

# Time and timezone handling
pytz>=2023.3

# Data processing
pandas>=2.0.3
numpy>=1.24.3

# API and web frameworks
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Configuration management
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
SQLAlchemy>=2.0.23
alembic>=1.13.0

# Monitoring and logging
prometheus-client>=0.19.0
structlog>=23.2.0

# Retry mechanisms
tenacity>=8.2.3

# Scheduling
APScheduler>=3.10.4
schedule>=1.2.0

# Testing
pytest>=7.4.3
pytest-mock>=3.12.0
pytest-cov>=4.1.0
requests-mock>=1.11.0

# Development tools
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1
isort>=5.12.0

# For Airflow (optional)
apache-airflow>=2.7.3
apache-airflow-providers-postgres>=5.6.0

# Additional utilities
click>=8.1.7
rich>=13.7.0